+++ commands/ipblock.go
@@ -152,8 +152,6 @@ Required values to run command:
 	deleteCmd.AddBoolFlag(config.ArgWaitForRequest, "", config.DefaultWait, "Wait for the Request for IpBlock deletion to be executed")
 	deleteCmd.AddIntFlag(config.ArgTimeout, "", config.DefaultTimeoutSeconds, "Timeout option for Request for IpBlock deletion [seconds]")
 
-	ipblockCmd.AddCommand(ipconsumers())
-
 	return ipblockCmd
 }
 
@@ -229,159 +227,6 @@ func RunIpBlockDelete(c *core.CommandConfig) error {
 		return err
 	}
 	return c.Printer.Print(getIpBlockPrint(resp, c, nil))
-}
-
-// IpBlock Consumer Commands
-
-func ipconsumers() *core.Command {
-	ctx := context.TODO()
-	resourceCmd := &core.Command{
-		Command: &cobra.Command{
-			Use:              "consumers",
-			Short:            "Ip Consumers Operations",
-			Long:             `The sub-command of ` + "`" + `ionosctl ipblock consumers` + "`" + ` allows you to list information about where each IP address from an IpBlock is being used.`,
-			TraverseChildren: true,
-		},
-	}
-	globalFlags := resourceCmd.GlobalFlags()
-	globalFlags.StringSlice(config.ArgCols, defaultIpConsumerCols, "Columns to be printed in the standard output")
-	_ = viper.BindPFlag(core.GetGlobalFlagName(resourceCmd.Name(), config.ArgCols), globalFlags.Lookup(config.ArgCols))
-	_ = resourceCmd.Command.RegisterFlagCompletionFunc(config.ArgCols, func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		return allIpConsumerCols, cobra.ShellCompDirectiveNoFileComp
-	})
-
-	/*
-		List Consumers Command
-	*/
-	listResources := core.NewCommand(ctx, resourceCmd, core.CommandBuilder{
-		Namespace:  "ipblock",
-		Resource:   "consumers",
-		Verb:       "list",
-		ShortDesc:  "List IpConsumers",
-		LongDesc:   "Use this command to get a list of Resources where each IP address from an IpBlock is being used.\n\nRequired values to run command:\n\n* IpBlock Id",
-		Example:    consumersListIpBlockExample,
-		PreCmdRun:  PreRunIpBlockId,
-		CmdRun:     RunIpConsumersList,
-		InitClient: true,
-	})
-	listResources.AddStringFlag(config.ArgIpBlockId, "", "", config.RequiredFlagIpBlockId)
-	_ = listResources.Command.RegisterFlagCompletionFunc(config.ArgIpBlockId, func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		return getIpBlocksIds(os.Stderr), cobra.ShellCompDirectiveNoFileComp
-	})
-
-	return resourceCmd
-}
-
-func RunIpConsumersList(c *core.CommandConfig) error {
-	ipsConsumers := make([]resources.IpConsumer, 0)
-	ipBlock, _, err := c.IpBlocks().Get(viper.GetString(core.GetFlagName(c.NS, config.ArgIpBlockId)))
-	if err != nil {
-		return err
-	}
-	if properties, ok := ipBlock.GetPropertiesOk(); ok && properties != nil {
-		if ipCons, ok := properties.GetIpConsumersOk(); ok && ipCons != nil {
-			for _, ip := range *ipCons {
-				ipsConsumers = append(ipsConsumers, resources.IpConsumer{IpConsumer: ip})
-			}
-		}
-	}
-	return c.Printer.Print(getIpConsumerPrint(c, ipsConsumers))
-}
-
-// Output Printing
-
-var (
-	defaultIpConsumerCols = []string{"Ip", "NicId", "ServerId", "DatacenterId", "K8sNodePoolId", "K8sClusterId"}
-	allIpConsumerCols     = []string{"Ip", "Mac", "NicId", "ServerId", "ServerName", "DatacenterId", "DatacenterName", "K8sNodePoolId", "K8sClusterId"}
-)
-
-type IpConsumerPrint struct {
-	Ip             string `json:"Ip,omitempty"`
-	Mac            string `json:"Mac,omitempty"`
-	NicId          string `json:"NicId,omitempty"`
-	ServerId       string `json:"ServerId,omitempty"`
-	ServerName     string `json:"ServerName,omitempty"`
-	DatacenterId   string `json:"DatacenterId,omitempty"`
-	DatacenterName string `json:"DatacenterName,omitempty"`
-	K8sNodePoolId  string `json:"K8sNodePoolId,omitempty"`
-	K8sClusterId   string `json:"K8sClusterId,omitempty"`
-}
-
-func getIpConsumerPrint(c *core.CommandConfig, groups []resources.IpConsumer) printer.Result {
-	r := printer.Result{}
-	if c != nil {
-		if groups != nil {
-			r.OutputJSON = groups
-			r.KeyValue = getIpConsumersKVMaps(groups)
-			r.Columns = getIpConsumerCols(core.GetGlobalFlagName(c.Resource, config.ArgCols), c.Printer.GetStderr())
-		}
-	}
-	return r
-}
-
-func getIpConsumerCols(flagName string, outErr io.Writer) []string {
-	if viper.IsSet(flagName) {
-		var groupCols []string
-		columnsMap := map[string]string{
-			"Ip":             "Ip",
-			"Mac":            "Mac",
-			"NicId":          "NicId",
-			"ServerId":       "ServerId",
-			"ServerName":     "ServerName",
-			"DatacenterId":   "DatacenterId",
-			"DatacenterName": "DatacenterName",
-			"K8sNodePoolId":  "K8sNodePoolId",
-			"K8sClusterId":   "K8sClusterId",
-		}
-		for _, k := range viper.GetStringSlice(flagName) {
-			col := columnsMap[k]
-			if col != "" {
-				groupCols = append(groupCols, col)
-			} else {
-				clierror.CheckError(errors.New("unknown column "+k), outErr)
-			}
-		}
-		return groupCols
-	} else {
-		return defaultIpConsumerCols
-	}
-}
-
-func getIpConsumersKVMaps(rs []resources.IpConsumer) []map[string]interface{} {
-	out := make([]map[string]interface{}, 0, len(rs))
-	for _, r := range rs {
-		var rPrint IpConsumerPrint
-		if ip, ok := r.GetIpOk(); ok && ip != nil {
-			rPrint.Ip = *ip
-		}
-		if mac, ok := r.GetMacOk(); ok && mac != nil {
-			rPrint.Mac = *mac
-		}
-		if nicId, ok := r.GetNicIdOk(); ok && nicId != nil {
-			rPrint.NicId = *nicId
-		}
-		if serverId, ok := r.GetServerIdOk(); ok && serverId != nil {
-			rPrint.ServerId = *serverId
-		}
-		if serverName, ok := r.GetServerNameOk(); ok && serverName != nil {
-			rPrint.ServerName = *serverName
-		}
-		if datacenterId, ok := r.GetDatacenterIdOk(); ok && datacenterId != nil {
-			rPrint.DatacenterId = *datacenterId
-		}
-		if datacenterName, ok := r.GetDatacenterNameOk(); ok && datacenterName != nil {
-			rPrint.DatacenterName = *datacenterName
-		}
-		if nodepoolId, ok := r.GetK8sNodePoolUuidOk(); ok && nodepoolId != nil {
-			rPrint.K8sNodePoolId = *nodepoolId
-		}
-		if clusterId, ok := r.GetK8sClusterUuidOk(); ok && clusterId != nil {
-			rPrint.K8sClusterId = *clusterId
-		}
-		o := structs.Map(rPrint)
-		out = append(out, o)
-	}
-	return out
 }
 
 // Output Printing
