+++ commands/root.go
@@ -156,6 +156,7 @@ func addCommands() {
 	rootCmd.AddCommand(nic())
 	rootCmd.AddCommand(loadBalancer())
 	rootCmd.AddCommand(ipblock())
+	rootCmd.AddCommand(ipconsumer())
 	rootCmd.AddCommand(request())
 	rootCmd.AddCommand(snapshot())
 	rootCmd.AddCommand(image())
